auth service implementation and interface separate
